requirements.txt
streamlit
llama-index-llms-gemini
youtube-transcript-api
scrapetube
